refactor(applicant): CreateApplicantService 인터페이스 제거
